% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_utilities.R
\name{add_agency_columns_to_df}
\alias{add_agency_columns_to_df}
\title{Adds columns to a gtfsr data frame with the agency id and name}
\usage{
add_agency_columns_to_df(gtfs_obj, dfname)
}
\arguments{
\item{gtfs_obj}{a list of standard gtfsr objects}

\item{dfname}{dataframe to return}
}
\value{
select data frame with the agency id and name
}
\description{
adds agency details to a gtfsr dataframe
}
