% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_utilities.R
\name{assign_frequencies_to_all_stops}
\alias{assign_frequencies_to_all_stops}
\title{Get stop frequency for buses based on mtc headway calculations}
\usage{
assign_frequencies_to_all_stops(gtfs_obj, time_start1, time_end1, service)
}
\arguments{
\item{gtfs_obj}{a standard gtfsr object}
}
\value{
gtfs_obj a gtfsr object with route level
frequencies (routes_df_frequency) as a dataframe
and a spatial stops_sf (stops_sf_frequency) with median frequencies
}
\description{
should take: start_time, end_time, gtfs_object, and some kind of scheduleing (array of days or "weekend")
}
