% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{bind_list_of_routes_spatial_dataframes}
\alias{bind_list_of_routes_spatial_dataframes}
\title{put a list of spatial dataframes (with agency_id as the list key) together into one}
\usage{
bind_list_of_routes_spatial_dataframes(l1)
}
\arguments{
\item{a}{list of routes sp class dataframes}
}
\value{
an sp class dataframe
}
\description{
put a list of spatial dataframes (with agency_id as the list key) together into one
}
