% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_deprecated.R
\name{both_directions_bool_check}
\alias{both_directions_bool_check}
\title{given a string vector, check whether both 0 and 1 are in it}
\usage{
both_directions_bool_check(direction_ids)
}
\arguments{
\item{string}{vector of}
}
\value{
logical vector
}
\description{
given a string vector, check whether both 0 and 1 are in it
}
