% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{count_trips}
\alias{count_trips}
\alias{count_trips}
\title{for a mega-GTFSr dataframe, count the number of trips a bus takes through a given stop within a given time period}
\usage{
count_trips(rt_df)

count_trips(rt_df)
}
\arguments{
\item{a}{mega-GTFSr dataframe}

\item{a}{mega-GTFSr dataframe}
}
\value{
a dataframe of stops with a "Trips" variable representing the count trips taken through each stop for a route within a given time frame

a dataframe of stops with a "Trips" variable representing the count trips taken through each stop for a route within a given time frame
}
\description{
for a mega-GTFSr dataframe, count the number of trips a bus takes through a given stop within a given time period

for a mega-GTFSr dataframe, count the number of trips a bus takes through a given stop within a given time period
}
