% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_utilities.R
\name{count_trips_for_service}
\alias{count_trips_for_service}
\title{Get a set of stops for a route}
\usage{
count_trips_for_service(g1)
}
\arguments{
\item{a}{gtfsr object}
}
\value{
count of service by id
}
\description{
Get a set of stops for a route
}
