% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_utilities.R
\name{filter_by_schedule}
\alias{filter_by_schedule}
\title{Get stop frequency for buses based on mtc headway calculations}
\usage{
filter_by_schedule(gtfs_df, service = "weekday")
}
\arguments{
\item{gtfs_df}{a mega df object made by join_all_gtfs_tables}

\item{service}{default to "weekday", can also use "weekend" currently}
}
\value{
a gtfs mega df object filtered
}
\description{
Get stop frequency for buses based on mtc headway calculations
}
