% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{filter_by_time}
\alias{filter_by_time}
\title{Get trips and headways based on the number of trips a set of buses on a route take in a specified time frame}
\usage{
filter_by_time(rt_df, time_start = "06:00:00", time_end = "09:59:00")
}
\arguments{
\item{a}{dataframe made by joining all the GTFS tables together}

\item{a}{start time filter hh:mm:ss}

\item{an}{end time filter hh:mm:ss}
}
\value{
a mega-GTFSr dataframe filtered to rows of interest
}
\description{
Get trips and headways based on the number of trips a set of buses on a route take in a specified time frame
}
