% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{fix_hour}
\alias{fix_hour}
\alias{fix_hour}
\title{Format GTFS Time strings as standard time string}
\usage{
fix_hour(x)

fix_hour(x)
}
\arguments{
\item{a}{GTFS Time string}

\item{a}{GTFS Time string}
}
\value{
Time string with no hours greater than 24

Time string with no hours greater than 24
}
\description{
Format GTFS Time strings as standard time string

Format GTFS Time strings as standard time string
}
