% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{flag_and_filter_peak_periods_by_time}
\alias{flag_and_filter_peak_periods_by_time}
\alias{flag_and_filter_peak_periods_by_time}
\title{Filter a mega-GTFSr dataframe to rows/stops that occur on all weekdays, are buses, and
have a stop_time between 2 time periods}
\usage{
flag_and_filter_peak_periods_by_time(mega_df, period)

flag_and_filter_peak_periods_by_time(mega_df, period)
}
\arguments{
\item{period}{\itemize{
\item "AM" or "PM"
}}

\item{a}{mega-GTFSr dataframe made with the join_all_gtfs_tables() function}

\item{a}{mega-GTFSr dataframe made with the join_all_gtfs_tables() function}

\item{period}{\itemize{
\item "AM" or "PM"
}}
}
\value{
a mega-GTFSr dataframe filtered to TPA peak periods and flagged as AM or PM peak

a mega-GTFSr dataframe filtered to TPA peak periods and flagged as AM or PM peak
}
\description{
Filter a mega-GTFSr dataframe to rows/stops that occur on all weekdays, are buses, and
have a stop_time between 2 time periods

Filter a mega-GTFSr dataframe to rows/stops that occur on all weekdays, are buses, and
have a stop_time between 2 time periods
}
