% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_deprecated.R
\name{get_geoms}
\alias{get_geoms}
\title{Return a spatial dataframe with the geometries of high frequency routes}
\usage{
get_geoms(route_id, gtfs_obj, weekday = TRUE, buffer = 402.336)
}
\arguments{
\item{gtfs_obj}{is a gtfsr list of gtfs dataframes}

\item{l1}{is a route name}
}
\value{
routes geometries as polygons, for weekend service
}
\description{
Return a spatial dataframe with the geometries of high frequency routes
}
