% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_utilities.R
\name{get_mtc_511_gtfs}
\alias{get_mtc_511_gtfs}
\title{Use the gtfsr import_gtfs function to import an MTC 511 api endpoint to a standarf gtfsr object}
\usage{
get_mtc_511_gtfs(privatecode, api_key)
}
\arguments{
\item{privatecode}{this is the shortcode used by 511 to refer to operators}
}
\value{
a gtfsr object
}
\description{
Use the gtfsr import_gtfs function to import an MTC 511 api endpoint to a standarf gtfsr object
}
