% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_utilities.R
\name{get_route_frequency}
\alias{get_route_frequency}
\title{Get stop frequency for buses aggregated up to routes}
\usage{
get_route_frequency(gtfs_obj, time_start1, time_end1, service = service)
}
\arguments{
\item{gtfs_obj}{a standard gtfsr object}

\item{service}{e.g. "weekend" or "saturday"}

\item{start_time, }{}

\item{end_time, }{}
}
\value{
route_headways a dataframe of route headways
}
\description{
should take:
}
