% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_deprecated.R
\name{get_route_stats}
\alias{get_route_stats}
\title{make high frequency routes df into just routes and directions df, for use in construction geoms by route and direction sldf}
\usage{
get_route_stats(df1)
}
\arguments{
\item{a}{dataframe made of am_routes and pm_routes}
}
\value{
a dataframe of routes by direction with headway stats for peak periods
}
\description{
make high frequency routes df into just routes and directions df, for use in construction geoms by route and direction sldf
}
