% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_deprecated.R
\name{get_route_stats_no_direction}
\alias{get_route_stats_no_direction}
\title{make high frequency routes df into just routes and directions df, for use in construction geoms by route and direction sldf}
\usage{
get_route_stats_no_direction(df1)
}
\arguments{
\item{a}{dataframe made of am_routes and pm_routes}
}
\value{
a dataframe of routes with headway stats for peak periods averaged over both directions
}
\description{
make high frequency routes df into just routes and directions df, for use in construction geoms by route and direction sldf
}
