% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_deprecated.R
\name{get_routes}
\alias{get_routes}
\title{for a mega-GTFSr dataframe, reduce it to just a listing of routes}
\usage{
get_routes(rt_df)
}
\arguments{
\item{a}{mega-GTFSr dataframe}
}
\value{
a dataframe of routes
}
\description{
for a mega-GTFSr dataframe, reduce it to just a listing of routes
}
