% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_deprecated.R
\name{get_routes_with_geoms_and_stats}
\alias{get_routes_with_geoms_and_stats}
\title{given am routes and pm routes, return a spatial dataframe with the routes and their stats}
\usage{
get_routes_with_geoms_and_stats(am_routes, pm_routes)
}
\value{
spatial dataframe (polygons) of routes
}
\description{
given am routes and pm routes, return a spatial dataframe with the routes and their stats
}
