% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{get_single_route_geom}
\alias{get_single_route_geom}
\title{Return the geometries for a route as single line}
\usage{
get_single_route_geom(x, hf_l)
}
\arguments{
\item{a}{list with route_id and direction id}

\item{a}{list output by get_hf_geoms}
}
\value{
linestring with an id for route and direction
}
\description{
Return the geometries for a route as single line
}
