% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gtfs.R
\name{gtfs_as_sf_routes}
\alias{gtfs_as_sf_routes}
\title{Get route features (sf) for a gtfsr object}
\usage{
gtfs_as_sf_routes(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{a standard gtfsr object}
}
\value{
gtfs_obj a gtfsr object with route shapes
}
\description{
Get route features (sf) for a gtfsr object
}
