% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gtfs.R
\name{gtfs_as_sf_routes_buffer}
\alias{gtfs_as_sf_routes_buffer}
\title{Get route buffer features (sf) for a gtfsr object}
\usage{
gtfs_as_sf_routes_buffer(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{a standard gtfsr object}
}
\value{
gtfs_obj a gtfsr object with route buffer shapes
}
\description{
Get route buffer features (sf) for a gtfsr object
}
