% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gtfs.R
\name{gtfs_as_sf_stops}
\alias{gtfs_as_sf_stops}
\title{Get stop points simple features (sf) for a gtfsr object}
\usage{
gtfs_as_sf_stops(gtfs_obj)
}
\arguments{
\item{gtfs_obj}{a standard gtfsr object}
}
\value{
gtfs_obj a gtfsr object with point shapes
}
\description{
Get stop points simple features (sf) for a gtfsr object
}
