\name{GTFSr-package}
\alias{GTFSr-package}
\alias{GTFSr}
\docType{package}
\title{A package which provides some basic functionality to load a General Transit Feed Spec (GTFS) feed
and perform some simple operations.}
\description{
This package provides some S4 classes for importing and manipulating data from a GTFS feed.
}
\details{
\tabular{ll}{
Package: \tab GTFS_r\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-04\cr
License: \tab No idea?\cr
Depends: \tab methods\cr
}


}
\author{
Colin Broderick with inspiration from ondrejivanic's code for analysing Sydney's new public transport timetable.
Maintainer: Complain to Colin Broderick <colin@cbroderick.me>

}
\references{
~~ Literature or other references for background information ~~
For source of some of the code see ondrejivanic's repository 131500: \url{https://github.com/ondrejivanic/131500}

GTFS Reference Guide: \url{https://developers.google.com/transit/gtfs/reference}
}

\keyword{ package }
\keyword{ gtfs }
\keyword{ transport }
