% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{is_in_both_directions}
\alias{is_in_both_directions}
\title{check for bidirectional routes}
\usage{
is_in_both_directions(df_rt_dr)
}
\arguments{
\item{dataframe}{with route_id and direction_id columns}
}
\value{
boolean/logical vector indicating whether the route goes in both directions
}
\description{
check for bidirectional routes
}
