% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_deprecated.R
\name{is_in_both_periods}
\alias{is_in_both_periods}
\title{check for dual period stops}
\usage{
is_in_both_periods(df_rt_dr)
}
\arguments{
\item{dataframe}{with stop_id and "peak" columns}
}
\value{
boolean/logical vector indicating whether the stop is in both am and pm periods
}
\description{
check for dual period stops
}
