% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_deprecated.R
\name{is_loop_route}
\alias{is_loop_route}
\title{string match check for loop routes}
\usage{
is_loop_route(headsign)
}
\arguments{
\item{headsign}{vector}
}
\value{
boolean/logical vector indicating whether its a loop
}
\description{
string match check for loop routes
}
