% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{join_all_gtfs_tables}
\alias{join_all_gtfs_tables}
\alias{join_all_gtfs_tables}
\title{Make a dataframe GTFS tables all joined together for route frequency calculations}
\usage{
join_all_gtfs_tables(g)

join_all_gtfs_tables(g)
}
\arguments{
\item{a}{GTFSr object for a given provider with routes, stops, stop_times, etc}

\item{a}{GTFSr object for a given provider with routes, stops, stop_times, etc}
}
\value{
a mega-GTFSr data frame with stops, stop_times, trips, calendar, and routes all joined

a mega-GTFSr data frame with stops, stop_times, trips, calendar, and routes all joined
}
\description{
Make a dataframe GTFS tables all joined together for route frequency calculations

Make a dataframe GTFS tables all joined together for route frequency calculations
}
