% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gtfs_realtime_proto.R
\name{load_gtfs_realtime_proto}
\alias{load_gtfs_realtime_proto}
\title{Load gtfs-realtime .proto file}
\usage{
load_gtfs_realtime_proto()
}
\description{
Loads Descriptors from the gtfs-realtime.proto file.
}
\details{
Runs upon package load. This should only be need to be run again if the Descriptor pool has been reset through \code{\link[RProtoBuf]{resetDescriptorPool}} or some other way.
}
