% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{make_arrival_hour_less_than_24}
\alias{make_arrival_hour_less_than_24}
\title{Make a dataframe GTFS arrival_time column into standard time variable}
\usage{
make_arrival_hour_less_than_24(df)
}
\arguments{
\item{dataframe}{containing a GTFS-style "arrival_time" column (time values at +24:00:00)}
}
\value{
dataframe containing a GTFS-style "arrival_time" column (no time values at +24:00:00)
}
\description{
Make a dataframe GTFS arrival_time column into standard time variable
}
