% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_utilities.R
\name{merge_gtfsr_dfs}
\alias{merge_gtfsr_dfs}
\title{Merge gtfsr data frames across gtfs objects}
\usage{
merge_gtfsr_dfs(gtfs_obj_list, dfname)
}
\arguments{
\item{gtfs_obj_list}{a list of standard gtfsr objects}

\item{dfname}{dataframe to return}
}
\value{
one gtfsr dataframe object
}
\description{
merges gtfsr objects
}
