% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb827_functions.R
\name{process_april_amendment_1}
\alias{process_april_amendment_1}
\title{Get stop frequency for buses based on mtc headway calculations}
\usage{
process_april_amendment_1(x)
}
\arguments{
\item{x}{a row from a csv describing mtc 511 data sources}
}
\value{
a spatial dataframe for april amendment 1, or an error message
}
\description{
Get stop frequency for buses based on mtc headway calculations
}
