% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtfsrealtime.R
\name{read_gtfsrealtime}
\alias{read_gtfsrealtime}
\title{Parse GTFS-realtime feeds}
\usage{
read_gtfsrealtime(api_request)
}
\arguments{
\item{api_request}{httr or httr2 GET call}
}
\value{
message of type 'transit_realtime.FeedMessage'
}
\description{
Takes a GTFS-realtime feed api call and translates the binary data to a \code{RProtoBuf} message.
}
\examples{
\dontrun{

# Cairns SunBus (api currently does not require registration)
sunbus_cairns_url <- "https://gtfsrt.api.translink.com.au/api/realtime/CNS/TripUpdates"

sunbus_cairns_api_call <- sunbus_cairns_url |>
  httr2::request() |>
  httr2::req_method("GET") |>
  httr2::req_perform()

sunbus_cairns_message <- sunbus_cairns_api_call |>
  read_gtfsrealtime()

sunbus_cairns_message
}
}
