% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{saturday_bus_service}
\alias{saturday_bus_service}
\title{Get all times a bus stops during saturday service for any service id}
\usage{
saturday_bus_service(gtfs_df)
}
\arguments{
\item{a}{dataframe made by joining all the GTFS tables together}
}
\value{
a mega-GTFSr dataframe filtered to saturday services
}
\description{
Get all times a bus stops during saturday service for any service id
}
