% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{stop_headway}
\alias{stop_headway}
\title{Get a \code{sf} dataframe for gtfs stops}
\usage{
stop_headway(a_stop_id, a_service_id, stop_times_df, trips_df,
  by_route = FALSE, the_units = "mins")
}
\arguments{
\item{gtfs_obj}{a valid GTFS object}

\item{stop_id}{a valid stop_id}

\item{service_id}{: a valid service_id from the calendar_df}

\item{units}{: lubridate types: default to 'mins'}
}
\value{
a vector of headways
}
\description{
Get a \code{sf} dataframe for gtfs stops
}
\examples{
a_stop_id <- gtfs_obj$stops_df[sample(nrow(gtfs_obj$stops_df), 1),]$stop_id
a_service_id <- gtfs_obj$calendar_df[sample(nrow(gtfs_obj$calendar_df), 1),]$service_id
some_headways <- stop_headway(gtfs_obj$stop_times_df,gtfs_obj$trips_df,a_stop_id,a_service_id)
summary(some_headways)
}
