% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{stop_times_to_headways}
\alias{stop_times_to_headways}
\title{Get a vector of headways at a stop for a stop_times_df}
\usage{
stop_times_to_headways(stop_times_df)
}
\arguments{
\item{stop_times_df}{stop_times for just 1 stop}
}
\value{
a vector of headways
}
\description{
Get a vector of headways at a stop for a stop_times_df
}
