% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{stops_headways}
\alias{stops_headways}
\title{Get a \code{sf} dataframe for gtfs stops}
\usage{
stops_headways(stop_ids, service_id, stop_times_df, trips_df)
}
\arguments{
\item{stop_ids}{a character vector of stop ids of interest}

\item{service_id}{a string, indicates which services (e.g. weekday) to use}

\item{trips_df}{a valid GTFS trips table}

\item{stop_times}{a valid GTFS stop times table}
}
\value{
a dataframe with stats for stops
}
\description{
Get a \code{sf} dataframe for gtfs stops
}
\examples{
some_stops <- gtfs_obj$stops_df[sample(nrow(gtfs_obj$stops_df), 40),]
some_stops_sf <- stops_df_as_sf(some_stops)
plot(some_stops_sf)
}
