% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{sunday_bus_service}
\alias{sunday_bus_service}
\title{Get all times a bus stops during sunday service for any service id}
\usage{
sunday_bus_service(gtfs_df)
}
\arguments{
\item{a}{dataframe made by joining all the GTFS tables together}
}
\value{
a mega-GTFSr dataframe filtered to sunday services
}
\description{
Get all times a bus stops during sunday service for any service id
}
