% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{sunday_service_ids}
\alias{sunday_service_ids}
\title{Get the service id's for sunday}
\usage{
sunday_service_ids(g1)
}
\arguments{
\item{a}{gtfsr object}
}
\value{
service_ids character vector
}
\description{
Get the service id's for sunday
}
