% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{weekday_service_ids}
\alias{weekday_service_ids}
\title{Get the service id's for weekdays}
\usage{
weekday_service_ids(g1)
}
\arguments{
\item{a}{gtfsr object}
}
\value{
service_ids for weekday trips
}
\description{
Get the service id's for weekdays
}
