% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_deprecated.R
\name{write_to_geopackage_with_date}
\alias{write_to_geopackage_with_date}
\title{write a spatial dataframe to the current working directory as a geopackage (with date in name-seconds since the epoch)}
\usage{
write_to_geopackage_with_date(spdf)
}
\arguments{
\item{spatial}{dataframe}
}
\value{
nothing
}
\description{
write a spatial dataframe to the current working directory as a geopackage (with date in name-seconds since the epoch)
}
