% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc_frequencies.R
\name{filter_by_time}
\alias{filter_by_time}
\alias{filter_by_time}
\title{Filter a mega-GTFSr dataframe to rows/stops that occur on all weekdays, are buses, and
have a stop_time between 2 time periods}
\usage{
filter_by_time(rt_df, start_filter, end_filter)

filter_by_time(rt_df, start_filter, end_filter)
}
\arguments{
\item{a}{dataframe made by joining all the GTFS tables together}

\item{a}{start time filter hh:mm:ss}

\item{an}{end time filter hh:mm:ss}

\item{a}{dataframe made by joining all the GTFS tables together}

\item{a}{start time filter hh:mm:ss}

\item{an}{end time filter hh:mm:ss}
}
\value{
a mega-GTFSr dataframe filtered to rows of interest

a mega-GTFSr dataframe filtered to rows of interest
}
\description{
Filter a mega-GTFSr dataframe to rows/stops that occur on all weekdays, are buses, and
have a stop_time between 2 time periods

Filter a mega-GTFSr dataframe to rows/stops that occur on all weekdays, are buses, and
have a stop_time between 2 time periods
}
